requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
pdf2image
python-dotenv
google-generativeai
python-dotenv
streamlit
pandas
requests